Updated readme and --help to reflect glob bug
After npm installing and running `npm run copyright:fix` with

"copyright:fix": "copyright src/**/*.js --fix",

in the `package.json`, it was only hitting `.js` files 1 folder deep.

This was only occuring when run as an npm task, not when run with

`node copyright.js some_glob --fix`.

After wrapping the glob in single quotes, it was hitting all of the
nested directories as well.